Island: Move auth resource utils to authentication_service
