+++ monkey/monkey_island/cc/services/authentication_service/flask_resources/register.py
@@ -6,9 +6,11 @@ from flask.typing import ResponseValue
 from flask_security.views import register
 
 from monkey_island.cc.flask_utils import AbstractResource
-from monkey_island.cc.resources.auth.credential_utils import get_username_password_from_request
 from monkey_island.cc.server_utils.response_utils import response_to_invalid_request
 from monkey_island.cc.services.authentication_service import AuthenticationService
+from monkey_island.cc.services.authentication_service.utils import (
+    get_username_password_from_request,
+)
 
 logger = logging.getLogger(__name__)
