+++ monkey/monkey_island/cc/services/authentication_service/utils.py
